Develop

+++ gui/src/app/pages/share-collection/share-collection.component.ts
@@ -4,10 +4,11 @@ import {CollectionService} from '../../Services/collection.service';
 import {Role} from '../../Entities/Role';
 import {RoleService} from '../../Services/role.service';
 import {ActivatedRoute} from '@angular/router';
-import {NbDialogService} from '@nebular/theme';
 import {ShareCollectionDialogComponent} from '../share-collection-dialog/share-collection-dialog.component';
 import {UserCollection} from '../../Entities/UserCollection';
 import {faTrash} from '@fortawesome/free-solid-svg-icons';
+import {DialogService} from "primeng/dynamicdialog";
+import {ConfirmationService} from "primeng/api";
 
 @Component({
   selector: 'app-share-collection',
@@ -25,7 +26,8 @@ export class ShareCollectionComponent implements OnInit {
   constructor(private collectionService: CollectionService,
               private roleService: RoleService,
               private route: ActivatedRoute,
-              private dialogService: NbDialogService) {
+              public dialogService: DialogService,
+              private confirmationService: ConfirmationService) {
   }
 
   ngOnInit() {
@@ -42,14 +44,24 @@ export class ShareCollectionComponent implements OnInit {
 
   openDialog() {
 
-    this.dialogService.open(ShareCollectionDialogComponent, {context: {collectionId: this.collectionId}}).onClose.subscribe(response => {
+    this.dialogService.open(ShareCollectionDialogComponent, {data: {collectionId: this.collectionId},
+          header: 'Share collection to',
+          height: '250px'})
+      .onClose
+      .subscribe(response => {
       this.collectionService.getUsers(this.collectionId).subscribe(data => this.userCollections = data);
     });
   }
 
   deleteUser(userId: number) {
-    this.collectionService.deleteUserFromCollection(userId, this.collectionId).subscribe(d =>
-      this.collectionService.getUsers(this.collectionId).subscribe(data => this.userCollections = data));
+    this.confirmationService.confirm({
+      message: 'Are you sure that you want to perform this action?',
+      accept: () => {
+        this.collectionService.deleteUserFromCollection(userId, this.collectionId).subscribe(d =>
+          this.collectionService.getUsers(this.collectionId).subscribe(data => this.userCollections = data));
+      }
+    });
+
   }
 
 }
